Add javaapi collection forwarders for 2.12 and 2.11
Adds 2.13 collection compatible forwarders for scala 2.11 and scala 2.12, see #346.

2.11 has converters in traits WrapAsJava and WrapAsScala, while 2.12 has them in  AsJavaConverters and AsScalaConverters.
The method names are the same, but unfortunately both CollectionConverters objects have to be separately defined for both sources, because of this import.

Tests are added in a Java file, to verify that the methods can be called as java api.