Scala 3.3.3, 2.13.13, 2.12.19
I didn't go to 2.13.14 because I'd have to upgrade Scala.js as
well and we seek to avoid doing that in this repo unless we
must, for maximum compatibility

going from Scala 3.2 to 3.3 might seem inconsistent with that,
but I think we should do it, because all of the messaging around
Scala 3 has been super consistent that 3.3 is the first LTS and
older versions are dead. plus 3.3 isn't at all new anymore.