update `Using` from Scala 2.13.17 and add `UsingTest`
fixes #692

includes Scala upgrades: Scala 2.13.17, 2.12.20 (was .13, .19),
Scala.js (because only latest Scala.js is available for 2.13.17,
as per Seb on the contributors forum)

copies the sources from scala/scala repo, reformatted using scalafmt

both `Using` and `UsingTest` needed slight adjustment to compile on
2.11 and 2.12; and `UsingTest` needed to be adjusted to use
scala.util.control.compat.ControlThrowable

I tried to get `UsingTest` passing on Scala 3. on the JVM, this was
sufficient to get the 2.13.17 stdlib so that the tests pass:
  dependencyOverrides += "org.scala-lang" % "scala-library" % scala213
but it doesn't do the trick on JS or Native for reasons unknown.
so I worked around by putting `UsingTest` under `src-jvm`.
once we are on a Scala 3.3 LTS version that has taken the 2.13.17
upgrade, it could be moved to regular `src`. (but anyway, who cares...)